Change vcmap implementation to template
Now it's possible to parametrize tbb concurrent hash map based engine
with allocator type.
